table, tablets: use partitioned_sstable_set to collect compaction groups into a table
With the existing compound_sstable_set, every sstable is considered when
reading a partition or when testing a partition for existence when merging
a memtable into cache. Since there can be ~100 tablets on a shard, this
results in CPU amplification of 100.

Fix by using partitioned_sstable_set, which filters out unrelated
sstables.

It would be better to take advantage of the fixed boundaries between
compaction groups, but this is a simple short-term fix that delivers
most of the advantages.