test/cql-pytest: tests for "date" type
This patch adds a few simple tests for the values of the "date" column
type, and how it can be initialized from string or integers, and what do
those values mean.

Two of the tests reproduce issue #17066, where validation is missing
for values that don't fit in a 32-bit unsigned integer.

Refs #17066

Signed-off-by: Nadav Har'El <nyh@scylladb.com>