test/cql-pytest: test_select_from_mutation_fragments.py: enable extra logging
Turn on DEBUG-level logging for the 'table' logger, for the tests in
this suite. These tests are very sensitive to things like flushing
memtables and although they have been very stable, recently we have seen
a failure in one of them, which is not possible to debug with the amount
of logs we have.

Refs: #17031