api: add /storage_service/describe_ring/{keyspace}/{table}
this API is added to support the tablet keyspace, which is a
per-table replica strategy, with which each table is composed
of multiple tablets, each of which is replicated on its own
replicas. so, in order to schedule repair, scylla manager needs
to understand the topology of keyspace on per-table basis instead
of on per-keyspace basis.

in this change, the new RESTful API of
`/storage_service/describe_ring/{keyspace}/{table}` enables us to
retrieve a list of token range and its corresponding endpoints with
given table.

Refs #16509

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>