doc: remove the OSS and Enterprise Features pages
This commit removes the following pages:
- ScyllaDB Open Source Features
- ScyllaDB Enterprise Features

They were outdated, incomplete, and misleading.
They were also redundant, as the per-release
updates are added as Release Notes.

With this update, the features listed on the removed
pages are added under the common page: ScyllaDB Features.

Note: No redirections are added, because no file paths
or URLs are changed with this commit.

Fixes https://github.com/scylladb/scylladb/issues/13485

Refs https://github.com/scylladb/scylladb/issues/16496