tools/scylla-nodetool.cc: handle API request failures gracefully
Currently, error handling is done via catching
http::unexpected_status_error and re-throwing an std::runtime_error.
Turns out this no longer works, because this error will only be thrown
by the http client, if the request had an expected reply code set.
The scylla_rest_client doesn't set an expected reply code, so this
exception was never thrown for some time now.
Furthermore, even when the above worked, it was not too user-friendly as
the error message only included the reply-code, but not the reply
itself.

So in this patch this is fixed:
* The handling of http::unexpected_status_error is removed, we don't
  want to use this mechanism, because it yields very terse error
  messages.
* Instead, the status code of the request is checked explicitely and all
  cases where it is not 200 are handled.
* A new api_request_failed exception is added, which is throw for all
  non-200 statuses with the extracted error message from the server (if
  any).
* This exception is caught by main, the error message is printed and
  scylla-nodetool returns with a new distinct error-code: 4.

With this, all cases where the request fails on ScyllaDB are handled and
we shouldn't hit cases where a nodetool command fails with some
obscure JSON parsing error, because the error reply has different JSON
schema than the expected happy-path reply.